backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiomysql==0.2.0
alembic==1.13.1
python-slugify==8.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-socketio==5.10.0
openai-whisper==20231117
TTS==0.22.0
transformers==4.36.0
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
celery==5.3.4
redis==5.0.1
# Resume parsing libraries
PyPDF2==3.0.1
python-docx==1.1.2
# OpenAI API for dynamic question generation
openai==1.12.0

